Implemented AMD based module and dependency loading.
Based on the AMD spec and CommonJS spec. Implemented via
piewpiew.require() and piewpiew.define(). These methods are both
attached to the piewpiew object, rather than as free or global variables
as per the AMD spec, just because they are not necessarily complete
implementations of require() and define(), so I'd rather not give any
false impressions that they are.